Rename config.hpp to common.hpp
