+++ src/core/common.hpp
